creating hook
