+++ src/components/UnifiedSidenav/UnifiedConstants.ts
@@ -1,0 +1,15 @@
+export interface BreadCrumb {
+    path: string;
+    title: string;
+}
+
+export interface TocItem {
+    label: string;
+    glyph?: string;
+    url?: string;
+    group?: boolean;
+    prefix?: string;
+    collapsible?: boolean;
+    breadcrumbs?: BreadCrumb[];
+    items?: TocItem[];
+}
